Added file via uploading
This is a quiz kind of program. Here, the user can play three quizzes, if the user succeeds first round then the user can play second and so the third one. If the user can't succeed current around then the user can play the next round but the user can't get any reward ;)